… .coverage
… .github/workflows/ci-uv.yml
… .github/workflows/test-ci-uv.yml
… CLAUDE.md
… CONTRIBUTING.md
… DEVELOPER-ONBOARDING.md
… MIGRATION-NOTICE.md
… README.md
… backup-requirements.txt
… cleanup-report.md
… environment-snapshot.txt
… migration-checklist.md
… performance-benchmarks.md
… pyproject.toml
… rollback-test-report.md
… scripts-backup/README.md
… scripts/env-setup.sh
… scripts/uv-helpers.sh
… specs/001-chat-application/quickstart.md
… src/dankerchat/__pycache__/__init__.cpython-312.pyc
… tests/__pycache__/test_dependency_installation.cpython-312-pytest-8.4.2.pyc
… tests/__pycache__/test_pyproject_creation.cpython-312-pytest-8.4.2.pyc
… tests/__pycache__/test_uv_commands.cpython-312-pytest-8.4.2.pyc
… tests/__pycache__/test_uv_installation.cpython-312-pytest-8.4.2.pyc
… tests/__pycache__/test_virtual_environment.cpython-312-pytest-8.4.2.pyc
… tests/test_dependency_installation.py
… tests/test_pyproject_creation.py
… tests/test_uv_commands.py
… tests/test_uv_installation.py
… tests/test_virtual_environment.py